Set travis to always send email on success.
We got into a weird state with our alarm due to not always sending
success emails. We were set up to send success emails only on change.
Our utils-common master is currently building successfully, but we're in
alarm. Here's what happened.

1. Cron failed on Sunday -> email sent, parsed by our alarm since it is
a cron.
2. Rerun succeeded on Monday -> email sent, not parsed by our alarm
since it is not a cron.
3. Cron succeeded ever since -> email not sent since we only send emails
if build status changes.

Effectively, without doing this we will be stuck in alarm until the
build fails and then succeeds again with under a cron build.